Basic, working layout of the game using WebSockets
* Removed many debug statements.
* Where game data is needed inside a template, the server will send a
fully filled in template using Jinja2. When no game data is needed for a
screen, the client simply unhides that portion of the screen. No page
refreshes anywhere.
* Moved jukejoint variables on Person into the Mind.
* Since the player may make multiple playthroughs and `songs` is a
static variable, songs are reset at the beginning of
establish_monologue() so that the player see's the entirety of the song.
* Modified Song.py in order to better suit the web versions needs.
* Added many new screens. Logic of showing/hiding is embedded inside the
html files.